Step 1: rename infertsconfig file to fix capitalization
