+++ src/inferTsconfig2.test.ts
