+++ src/inferTsconfig2.ts
